move app component to components dir
